Moved data flow tests into data flow module + scalafmt (sigh).
