+++ dataflowengine/build.sbt
@@ -1,3 +1,8 @@
 name := "dataflowengine"
 
 dependsOn(Projects.semanticcpg)
+
+libraryDependencies ++= Seq(
+  "org.scalatest" %% "scalatest" % "3.0.3" % Test,
+  "io.shiftleft" %% "fuzzyc2cpg" % Versions.fuzzyc2cpg
+)
